requirements.txt
pandas
numpy
xgboost
scikit-learn
shap
google-generativeai
python-dotenv
matplotlib
requests